submit java-oop-ru/interfaces

+++ java-oop-ru/interfaces/src/main/java/exercise/Cottage.java
@@ -3,16 +3,16 @@ package exercise;
 // BEGIN
 public class Cottage implements Home {
 
-    private int area;
+    private double area;
     private int floorCount;
 
-    public Cottage(int area, int floorCount) {
+    public Cottage(double area, int floorCount) {
         this.area = area;
         this.floorCount = floorCount;
     }
 
     @Override
-    public int getArea() {
+    public double getArea() {
         return area;
     }
